rasdaemon: ras-events: Fix event's file endianess
tep_set_file_bigendian() is never called so for big-endian hosts, the
libtraceevents thinks it's handling a little-endian file on a big-endian
architecture and swap the bytes when accessing 16-bits or 32-bits words.
It leads to tep_print_event failing because the event ID isn't
recognized.

Call tep_set_file_bigendian() before tep_print_event() to make the
libtraceevent aware of the file's endianess endianess.
Use the ENDIAN macro to state the endianess as this macro is not used
elsewhere.

Signed-off-by: Bastien Curutchet <bastien.curutchet@bootlin.com>